+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # markdown-for-agents
 
+## 0.1.3
+
+### Patch Changes
+
+- add etag, vary accept header support
+
 ## 0.1.2
 
 ### Patch Changes