+++ packages/middleware/fastify/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @markdown-for-agents/fastify
 
+## 0.1.3
+
+### Patch Changes
+
+- add etag, vary accept header support
+
+- Updated dependencies []:
+    - markdown-for-agents@0.1.3
+
 ## 0.1.2
 
 ### Patch Changes